refa: refactor __main__ import
