+++ src/pyinnodb/cli/__init__.py
@@ -1,0 +1,5 @@
+from .main import *
+from . import iter_record
+from . import static_usage
+from . import sql
+from . import validate
