+++ src/pyinnodb/cli/main.py
@@ -1,0 +1,19 @@
+import click
+import logging
+from pyinnodb.disk_struct.fsp import MFspPage
+from pyinnodb import const
+from io import BytesIO
+
+
+@click.group()
+@click.argument("fn", type=click.File("rb"))
+@click.option(
+    "--log-level", type=click.Choice(["DEBUG", "ERROR", "INFO"]), default="ERROR"
+)
+@click.pass_context
+def main(ctx, fn, log_level):
+    logging.basicConfig(level=log_level)
+    ctx.ensure_object(dict)
+    ctx.obj["fn"] = fn
+    fsp_page = MFspPage.parse_stream(fn)
+    ctx.obj["fsp_page"] = fsp_page
